refactor: seperate widgets
