+++ lib/widgets/drawer_menu.dart
@@ -1,0 +1,68 @@
+import 'package:flutter/material.dart';
+
+// 드로어 메뉴 위젯
+class DrawerMenu extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Drawer(
+      child: ListView(
+        padding: EdgeInsets.zero,
+        children: <Widget>[
+          DrawerHeader(
+            decoration: BoxDecoration(
+              color: Color(0xFF4A7DFF),
+            ),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: [
+                Container(
+                  width: 60,
+                  height: 60,
+                  decoration: BoxDecoration(
+                    shape: BoxShape.circle,
+                    // image: DecorationImage(
+                    //   image: AssetImage(
+                    //       'assets/profile_image.png'), // 프로필 이미지
+                    //   fit: BoxFit.fill,
+                    // ),
+                  ),
+                ),
+                SizedBox(height: 10),
+                Text(
+                  '사용자 프로필',
+                  style: TextStyle(
+                    fontSize: 18,
+                    color: Colors.white,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+              ],
+            ),
+          ),
+          ListTile(
+            leading: Icon(Icons.home),
+            title: Text('홈'),
+            onTap: () {
+              Navigator.pushNamed(context, '/');
+            },
+          ),
+          ListTile(
+            leading: Icon(Icons.pageview),
+            title: Text('페이지 1'),
+            onTap: () {
+              Navigator.pushNamed(context, '/page1');
+            },
+          ),
+          ListTile(
+            leading: Icon(Icons.pageview),
+            title: Text('페이지 2'),
+            onTap: () {
+              Navigator.pushNamed(context, '/page2');
+            },
+          ),
+        ],
+      ),
+    );
+  }
+}
